Flag read posts by what's on screen, batch UI updates for usability
What "read" meant exactly was proving a little awkward. On the one hand,
users probably expect that leafing through pages would flag them as read
so they wouldn't see the same posts again after refreshes/changing users.
On the other hand, new posts are bumped to the top, and if each page
turn updated this list, you would start missing posts because the order
wouldn't stay consistent. Further, as posts are coming in they will
sometimes populate the page temporarily -- we want this in cases where
network speed is slow or the rate limiting turns out to be harsh, so it
isn't all-or-nothing and you get at least a _sense_ of progress. Plus,
the ordering of states coming in could mark some pages as read when they
were barely even rendered at all.

So, we're going to separate the UI flushing from the state updates in
componentDidUpdate, only flushing the UI when changing users and only if
nothing is being pulled at the moment. This required some improved efficiency
of the user_favs update system, to update per user so that flushing the UI
at all wouldn't be prohibitively expensive (i.e. reloading and reparsing the
entire database). Also, this fixes iterating to get more than one page, which
previously didn't actually work.