Implement sort by, style all the things
Sort by newest first by default. After using the UI seriously for a bit,
I realized that "unviewed first" is a bit unintuitive, because things
jump around a lot as you jump between users without clearing the whole
list of unviewed posts. This workflow feels better to me: to browse
newest-first for the most part, and jump to unviewed occasionally to
clear out some of the posts I missed.